requirements/local.txt
-r common.txt
django-debug-toolbar==1.6
Werkzeug==0.11.15
ipython==5.2.2
